after going to the shop we noticed that Double wasnt a good way to handle money. Usability should be kinda better too

+++ basket.playground/Sources/Shoppable.swift
@@ -3,7 +3,7 @@ import Foundation
 public protocol Shoppable: CustomStringConvertible {
     var name: String { get }
     var symbol: String { get }
-    var price: Double { get }
+    var price: Decimal { get }
 }
 
 /*
